Fix return type of setActualStock

+++ src/comp1110/exam/Q3SimpleStockManager.java
@@ -84,7 +84,7 @@ public class Q3SimpleStockManager {
      * @param actual The quantity actually in the store
      * @return The stock loss or gain (new current – old current)
      */
-    public double setActualStock(String sku, int actual) {
+    public int setActualStock(String sku, int actual) {
         return 0;  // FIXME Question 3ih: complete this function
     }
